Merge bugfix/ubsan-cleanups to release/v3.2.0

* commit 'cfc6e4060201d411113bbeeccf6e40fa47c4f4fa':
  Address UBSan review findings
  Fix UBSan CI build
  Run CI with undefined behavior sanitizer
  Fix undefined behavior sanitizer issues

+++ src/parsers/bits.c
@@ -19,11 +19,10 @@ static HParseResult *parse_bits(void *env, HParseState *state) {
      * requested bit count; callers that need wider fields should use h_bytes()
      * or compose smaller parsers.
      */
-    // h_read_bits takes int; cast is required by its signature
     if (env_->signedp)
-        result->token_data.sint = h_read_bits(&state->input_stream, (int)env_->length, true);
+        result->token_data.sint = h_read_bits(&state->input_stream, env_->length, true);
     else
-        result->token_data.uint = h_read_bits(&state->input_stream, (int)env_->length, false);
+        result->token_data.uint = h_read_bits(&state->input_stream, env_->length, false);
     result->index = 0;
     result->bit_length = 0;
     result->bit_offset = 0;
